+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * Use new import and export window for profile import, export, and direct player share
 * Improve warning shown when the self-cast modifier is enabled
 * Add `bonusbar` and `nobonusbar` macro conditions
+  * For Dragonriding, use bonus bar 5, for Dracthyr Soar use bonus bar 1
 
 ### Fixed